requirements.txt
Flask==2.3.3
gunicorn==21.2.0
zeep==4.2.1
lxml==4.9.3
xmlsec==1.3.13
pyOpenSSL==23.2.0
pytz==2023.3
python-dotenv==1.0.0